+++ lib/twenv/commands/write_tweet.rb
@@ -6,7 +6,7 @@ class TWEnv::WriteTweet < TWEnv::Command
 
   def process
     raise Pry::CommandError, "$EDITOR is not set" if empty?(ENV['EDITOR'])
-    file = Tempfile.new 'twenv', command_storage_path
+    file = Tempfile.new 'twenv', storage_path
     system ENV['EDITOR'], file.path
     tweet = file.read
     raise Pry::CommandError, "tweet content is empty" if empty?(tweet)